Initial iteration of ECS system with rendering system for Sprite and Tiles (including animated Tiles) (#1)
Early implementation of ECS system along with a rendering system for Sprite and Tiles (including animated Tiles). The implementation of ECS is still subjects to change. While the RenderingSystem is still lack many features.

But this build already covers sprite animation and tile rendering. Making entities and custom script is also now possible.