Tied camera scale with scene. And tied camera to screen. (#5)
Scaling towards the screen size should now only be done with renderContext->camera_size. Camera will be tied with screen, meaning it will be wrap to fit the screen. So zoom effect can be done by lowering the game->renderContext->camera_size.

In the future, the engine will by default render to a texture, and scaling camera render with the screen can happen in separate process than with entity rendering.